requirements.txt
bcrypt==3.1.6
bleach==3.1.0
blinker==1.4
cffi==1.11.5
Click==7.0
decorator==4.3.2
elasticsearch==6.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
pycparser==2.19
six==1.12.0
SQLAlchemy==1.2.17
SQLAlchemy-Utils==0.33.11
urllib3==1.24.1
validators==0.12.4
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-Alchemy==0.16.8
WTForms-Components==0.10.3
WTForms-SQLAlchemy==0.1
